+++ src/utils/close_askReason.ts
@@ -1,5 +1,3 @@
-const Discord = require("discord.js");
-
 /*
 Copyright 2023 Sayrix (github.com/Sayrix)
 
@@ -16,33 +14,34 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-module.exports = {
-	async closeAskReason(interaction, client) {
-		if (
-			client.config.whoCanCloseTicket === "STAFFONLY" &&
-			!interaction.member.roles.cache.some((r) => client.config.rolesWhoHaveAccessToTheTickets.includes(r.id))
-		)
-			return interaction
-				.reply({
-					content: client.locales.ticketOnlyClosableByStaff,
-					ephemeral: true,
-				})
-				.catch((e) => console.log(e));
+import { ButtonInteraction, GuildMember } from "discord.js";
+import { DiscordClient } from "../Types";
 
-		const modal = new Discord.ModalBuilder().setCustomId("askReasonClose").setTitle(client.locales.modals.reasonTicketClose.title);
+export const closeAskReason = async(interaction: ButtonInteraction, client: DiscordClient) => {
+	if (
+		client.config.whoCanCloseTicket === "STAFFONLY" &&
+		!(interaction.member as GuildMember | null)?.roles.cache.some((r) => client.config.rolesWhoHaveAccessToTheTickets.includes(r.id))
+	)
+		return interaction
+			.reply({
+				content: client.locales.ticketOnlyClosableByStaff,
+				ephemeral: true,
+			})
+			.catch((e) => console.log(e));
 
-		const input = new Discord.TextInputBuilder()
-			.setCustomId("reason")
-			.setLabel(client.locales.modals.reasonTicketClose.label)
-			.setStyle(Discord.TextInputStyle.Paragraph)
-			.setPlaceholder(client.locales.modals.reasonTicketClose.placeholder)
-			.setMaxLength(256);
+	const modal = new Discord.ModalBuilder().setCustomId("askReasonClose").setTitle(client.locales.modals.reasonTicketClose.title);
 
-		const firstActionRow = new Discord.ActionRowBuilder().addComponents(input);
-		modal.addComponents(firstActionRow);
-		await interaction.showModal(modal).catch((e) => console.log(e));
-	},
-};
+	const input = new Discord.TextInputBuilder()
+		.setCustomId("reason")
+		.setLabel(client.locales.modals.reasonTicketClose.label)
+		.setStyle(Discord.TextInputStyle.Paragraph)
+		.setPlaceholder(client.locales.modals.reasonTicketClose.placeholder)
+		.setMaxLength(256);
+
+	const firstActionRow = new Discord.ActionRowBuilder().addComponents(input);
+	modal.addComponents(firstActionRow);
+	await interaction.showModal(modal).catch((e) => console.log(e));
+}
 
 /*
 Copyright 2023 Sayrix (github.com/Sayrix)
