+++ src/utils/test.ts
@@ -1,0 +1,18 @@
+async function waiter(a: string) {
+    return {
+        kc: "Aa"
+    };
+}
+
+async function waiters(a?: string) {
+    return a;
+}
+
+async function test() {
+    let a = await waiters();
+    if(!a) return;
+    while(true) {
+        const k = await waiter(a);
+        a = k.kc;
+    }
+}
